refactor: replace Cart API component
